Improve wishlist feature
